Rename `en` to `en_US`
