+++ test/unit/views/payment-options-view.js
@@ -4,7 +4,7 @@ var BaseView = require('../../../src/views/base-view');
 var CardView = require('../../../src/views/payment-sheet-views/card-view');
 var PaymentOptionsView = require('../../../src/views/payment-options-view');
 var DropinModel = require('../../../src/dropin-model');
-var strings = require('../../../src/translations/en');
+var strings = require('../../../src/translations/en_US');
 var fake = require('../../helpers/fake');
 var fs = require('fs');
 var analytics = require('../../../src/lib/analytics');
